pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Paul Nikolaevskiy <paulnikolaevsky@gmail.com>"]
packages = [
	{ include = "engine" },
]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^7.1.3"
pytest-cov = "^3.0.0"
coverage = "^6.4.4"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
ipykernel = "^6.16.0"

[tool.poetry.scripts]
gendiff = "engine.scripts.gendiff:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
